[#64] feat: options, route 분리
