+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,23 +1,6 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-import KakaoProvider from 'next-auth/providers/kakao';
-
-const authOptions: NextAuthOptions = {
-  providers: [
-    KakaoProvider({
-      clientId: process.env.KAKAO_CLIENT_ID || '',
-      clientSecret: process.env.KAKAO_CLIENT_SECRET || '',
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID || '',
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || '',
-    }),
-  ],
-  pages: {
-    signIn: '/signin',
-  },
-};
+import authOptions from '@/app/api/auth/[...nextauth]/options';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
 
-export { handler as GET, handler as POST, authOptions };
+export { handler as GET, handler as POST };
